+++ source/dotnet/Services/Application/ProcessResult/IProcessActorResultApplicationService.cs
@@ -12,18 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
-using Energinet.DataHub.Wholesale.Domain.GridAreaAggregate;
+using Energinet.DataHub.Wholesale.Contracts;
 
-namespace Energinet.DataHub.Wholesale.Domain.ProcessOutput;
+namespace Energinet.DataHub.Wholesale.Application.ProcessResult;
 
-public interface IProcessOutputRepository
+/// <summary>
+/// This class provides the ability to retrieve a calculated result for a given step for a batch.
+/// </summary>
+public interface IProcessActorResultApplicationService
 {
-    /// <summary>
-    /// Create zip archives for each process in the batch.
-    /// The archive contains the basis data files and the result file.
-    /// </summary>
-    Task CreateBasisDataZipAsync(Batch completedBatch);
-
-    Task<Stream> GetZippedBasisDataStreamAsync(Batch batch);
+    Task<ProcessStepResultDto> GetResultAsync(ProcessStepResultRequestDto processStepResultRequestDto);
 }
