+++ source/dotnet/Services/Tests/Domain/ProcessActorResultAggregate/ProcessActorResultAggregateTests.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Wholesale.Domain.ProcessActorResultAggregate;
 using Energinet.DataHub.Wholesale.Domain.ProcessOutput;
 using FluentAssertions;
 using Xunit;
