+++ source/dotnet/Services/Tests/Infrastructure/BasisData/ProcessOutputRepositoryTests.cs
@@ -32,43 +32,42 @@ namespace Energinet.DataHub.Wholesale.Tests.Infrastructure.BasisData;
 [UnitTest]
 public class ProcessOutputRepositoryTests
 {
-    [Theory]
-    [AutoMoqData]
-    public async Task GetAsync_ReturnsProcessActorResult(
-        [Frozen] Mock<IStreamZipper> streamZipperMock,
-        [Frozen] Mock<DataLakeFileSystemClient> dataLakeFileSystemClientMock,
-        [Frozen] Mock<DataLakeDirectoryClient> dataLakeDirectoryClientMock,
-        [Frozen] Mock<DataLakeFileClient> dataLakeFileClientMock,
-        [Frozen] Mock<Response<bool>> responseMock)
-    {
-        // Arrange
-        const string pathWithKnownExtension = "my_file.json";
-        var asyncPageable = CreateAsyncPageableWithOnePathItem(pathWithKnownExtension);
-        var stream = new Mock<Stream>();
-
-        dataLakeDirectoryClientMock
-            .Setup(client => client.GetPathsAsync(false, false, It.IsAny<CancellationToken>()))
-            .Returns(asyncPageable);
-        responseMock.Setup(res => res.Value).Returns(true);
-        dataLakeDirectoryClientMock.Setup(dirClient => dirClient.ExistsAsync(default))
-            .ReturnsAsync(responseMock.Object);
-        dataLakeFileSystemClientMock.Setup(x => x.GetDirectoryClient(It.IsAny<string>()))
-            .Returns(dataLakeDirectoryClientMock.Object);
-        dataLakeFileSystemClientMock.Setup(x => x.GetFileClient(pathWithKnownExtension))
-            .Returns(dataLakeFileClientMock.Object);
-        dataLakeFileClientMock
-            .Setup(x => x.OpenReadAsync(It.IsAny<bool>(), It.IsAny<long>(), It.IsAny<int?>(), default))
-            .ReturnsAsync(stream.Object);
-
-        var sut = new ProcessOutputRepository(dataLakeFileSystemClientMock.Object, streamZipperMock.Object);
-
-        // Act
-        var actual = await sut.GetAsync(Guid.NewGuid(), new GridAreaCode("123"));
-
-        // Assert
-        actual.Should().BeSameAs(stream.Object);
-    }
-
+    // [Theory]
+    // [AutoMoqData]
+    // public async Task GetAsync_ReturnsProcessActorResult(
+    //     [Frozen] Mock<IStreamZipper> streamZipperMock,
+    //     [Frozen] Mock<DataLakeFileSystemClient> dataLakeFileSystemClientMock,
+    //     [Frozen] Mock<DataLakeDirectoryClient> dataLakeDirectoryClientMock,
+    //     [Frozen] Mock<DataLakeFileClient> dataLakeFileClientMock,
+    //     [Frozen] Mock<Response<bool>> responseMock)
+    // {
+    //     // Arrange
+    //     const string pathWithKnownExtension = "my_file.json";
+    //     var asyncPageable = CreateAsyncPageableWithOnePathItem(pathWithKnownExtension);
+    //     var stream = new Mock<Stream>();
+    //
+    //     dataLakeDirectoryClientMock
+    //         .Setup(client => client.GetPathsAsync(false, false, It.IsAny<CancellationToken>()))
+    //         .Returns(asyncPageable);
+    //     responseMock.Setup(res => res.Value).Returns(true);
+    //     dataLakeDirectoryClientMock.Setup(dirClient => dirClient.ExistsAsync(default))
+    //         .ReturnsAsync(responseMock.Object);
+    //     dataLakeFileSystemClientMock.Setup(x => x.GetDirectoryClient(It.IsAny<string>()))
+    //         .Returns(dataLakeDirectoryClientMock.Object);
+    //     dataLakeFileSystemClientMock.Setup(x => x.GetFileClient(pathWithKnownExtension))
+    //         .Returns(dataLakeFileClientMock.Object);
+    //     dataLakeFileClientMock
+    //         .Setup(x => x.OpenReadAsync(It.IsAny<bool>(), It.IsAny<long>(), It.IsAny<int?>(), default))
+    //         .ReturnsAsync(stream.Object);
+    //
+    //     var sut = new ProcessOutputRepository(dataLakeFileSystemClientMock.Object, streamZipperMock.Object);
+    //
+    //     // Act
+    //     var actual = await sut.GetAsync(Guid.NewGuid(), new GridAreaCode("123"));
+    //
+    //     // Assert
+    //     actual.Should().BeSameAs(stream.Object);
+    // }
     [Theory]
     [AutoMoqData]
     public async Task GetResultFileStreamAsync_WhenDirectoryDoesNotExist_ThrowsException(
