+++ source/dotnet/Services/WebApi/Configuration/ServiceCollectionExtensions.cs
@@ -22,6 +22,7 @@ using Energinet.DataHub.Wholesale.Application.JobRunner;
 using Energinet.DataHub.Wholesale.Application.Processes;
 using Energinet.DataHub.Wholesale.Application.ProcessResult;
 using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
+using Energinet.DataHub.Wholesale.Domain.ProcessActorResultAggregate;
 using Energinet.DataHub.Wholesale.Domain.ProcessOutput;
 using Energinet.DataHub.Wholesale.Infrastructure.BasisData;
 using Energinet.DataHub.Wholesale.Infrastructure.Core;
@@ -82,8 +83,9 @@ internal static class ServiceCollectionExtensions
         services.AddScoped<IProcessCompletedPublisher>(_ => null!); // Unused in the use cases of this app
         services.AddScoped<ICalculatorJobRunner>(_ => null!); // Unused in the use cases of this app
         services.AddScoped<ICalculatorJobParametersFactory>(_ => null!); // Unused in the use cases of this app
-        services.AddScoped<IProcessStepResultApplicationService, ProcessStepResultApplicationService>();
+        services.AddScoped<IProcessActorResultApplicationService, ProcessActorResultApplicationService>();
         services.AddScoped<IProcessActorResultMapper, ProcessActorResultMapper>();
+        services.AddScoped<IProcessActorResultRepository, ProcessOutputRepository>();
         services.AddScoped<IBatchRequestDtoValidator, BatchRequestDtoValidator>();
 
         services.ConfigureDateTime();
