+++ source/dotnet/Services/WebApi/Controllers/V2/ProcessStepResultController.cs
@@ -25,11 +25,11 @@ namespace Energinet.DataHub.Wholesale.WebApi.Controllers.V2;
 public class ProcessStepResultController : ControllerBase
 {
     private const string Version = "2.0";
-    private readonly IProcessStepResultApplicationService _processStepResultApplicationService;
+    private readonly IProcessActorResultApplicationService _processActorResultApplicationService;
 
-    public ProcessStepResultController(IProcessStepResultApplicationService processStepResultApplicationService)
+    public ProcessStepResultController(IProcessActorResultApplicationService processActorResultApplicationService)
     {
-        _processStepResultApplicationService = processStepResultApplicationService;
+        _processActorResultApplicationService = processActorResultApplicationService;
     }
 
     /// <summary>
@@ -41,7 +41,7 @@ public class ProcessStepResultController : ControllerBase
     [MapToApiVersion("2.1")]
     public async Task<IActionResult> GetAsync([FromBody] ProcessStepResultRequestDto processStepResultRequestDto)
     {
-        var resultDto = await _processStepResultApplicationService.GetResultAsync(processStepResultRequestDto).ConfigureAwait(false);
+        var resultDto = await _processActorResultApplicationService.GetResultAsync(processStepResultRequestDto).ConfigureAwait(false);
         return Ok(resultDto);
     }
 }
